+++ JrBoostLib/LogitBoostOptions.h
@@ -9,12 +9,12 @@ public:
     virtual ~LogitBoostOptions() = default;
 
     size_t iterationCount() const;
-    float eta() const;
+    double eta() const;
     bool highPrecision() const;
     AbstractOptions* baseOptions() const;
 
     void setIterationCount(size_t n);
-    void setEta(float eta);
+    void setEta(double eta);
     void setHighPrecision(bool b);
     void setBaseOptions(const AbstractOptions& opt);
 
@@ -25,7 +25,7 @@ private:
     LogitBoostOptions(const LogitBoostOptions&);
 
     size_t iterationCount_{ 100 };
-    float eta_{ 0.3f };
+    double eta_{ 0.3f };
     bool highPrecision_{ true };
     unique_ptr<AbstractOptions> baseOptions_;
 };
