+++ JrBoostLib/LogitBoostTrainer.h
@@ -11,8 +11,8 @@ public:
     virtual ~LogitBoostTrainer() = default;
 
     virtual void setInData(CRefXXf inData);
-    virtual void setOutData(const ArrayXf& outData);
-    virtual void setWeights(const ArrayXf& weights);
+    virtual void setOutData(const ArrayXd& outData);
+    virtual void setWeights(const ArrayXd& weights);
     virtual void setOptions(const AbstractOptions& opt);
 
     virtual BoostPredictor* train() const;
@@ -21,8 +21,8 @@ private:
     template<typename F>
     BoostPredictor* trainImpl_() const;
 
-    ArrayXf outData_;
-    ArrayXf weights_;
+    ArrayXd outData_;
+    ArrayXd weights_;
     unique_ptr<LogitBoostOptions> options_;
 
     CRefXXf inData_{ dummyArrayXXf };
